# api-swagger-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "viko.ramayudha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
...
        "/trainer": {
            "get": {
                "tags": [
                    "Trainer"
                ],
                "summary": "Trainer - Get All",
                "description": "Mengambil Data Trainer",
                "operationId": "listtrainer",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Trainer",
                                    "data": [
                                        {
                                            "id": "1",
                                            "nama": "Mukhamad Viko Ramayudha",
                                            "email": "ramaydha@dummy.com",
                                            "password": "******",
                                            "no_hp": "62813320200",
                                            "status": "1"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trainer/{id}": {
            "get": {
                "tags": [
                    "Trainer"
                ],
                "summary": "Trainer - Get By Id",
                "description": "Mengambil Data Trainer Berdasarkan ID",
                "operationId": "listTrainerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Masukkan ID Trainer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Trainer",
                                    "data": {
                                        "id": "1",
                                        "nama": "Mukhamad Viko Ramayudha",
                                        "email": "ramaydha@dummy.com",
                                        "password": "******",
                                        "no_hp": "62813320200",
                                        "status": "1"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data Trainer tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Data Trainer tidak ditemukan"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trainer/create": {
            "post": {
                "tags": [
                    "Trainer"
                ],
                "summary": "Trainer - Insert",
                "description": "Menambahkan Data Trainer",
                "operationId": "insertTrainer",
                "requestBody": {
                    "description": "Data Trainer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "nama": "Mukhamad Viko Ramayudha",
                                "email": "ramaydha@dummy.com",
                                "password": "******",
                                "no_hp": "62813320200",
                                "status": "1"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil menambahkan Data Trainer",
                                    "data": {
                                        "id": "",
                                        "nama": "Mukhamad Viko Ramayudha",
                                        "email": "ramaydha@dummy.com",
                                        "password": "******",
                                        "no_hp": "62813320200",
                                        "status": "1"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": [
                                        "The mhs nim field is required."
                                    ],
                                    "nama": [
                                        "The mhs nama field is required."
                                    ],
                                    "email": [
                                        "The mhs nama field is required."
                                    ],
                                    "password": [
                                        "The mhs nama field is required."
                                    ],
                                    "no_hp": [
                                        "The mhs nama field is required."
                                    ],
                                    "status": [
                                        "The mhs nama field is required."
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trainer/update": {
            "put": {
                "tags": [
                    "Trainer"
                ],
                "summary": "Trainer Update",
                "description": "Update data Trainer",
                "operationId": "updateTrainer",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "1",
                                "nama": "Mukhamad Viko Ramayudha",
                                "email": "ramaydha@dummy.com",
                                "password": "******",
                                "no_hp": "62813320200",
                                "status": "1"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil diubah"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/trainer/delete": {
            "delete": {
                "tags": [
                    "Trainer"
                ],
                "summary": "Trainer Delete",
                "description": "Hapus data Trainer",
                "operationId": "deleteTrainer",
                "requestBody": {
                    "description": "ID Trainer yang ingin dihapus",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "id": "1"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Data berhasil dihapus"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    }
}